Extract remote read contracts
